requirements.txt
asgiref==3.7.2
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
certifi==2024.2.2
cffi==1.17.1
channels==4.0.0
channels-redis==4.2.0
charset-normalizer==3.3.2
coloredlogs==15.0.1
constantly==23.10.4
crispy-bootstrap5==2023.10
cryptography==42.0.2
daphne==4.0.0
dj-database-url==2.1.0
Django==5.0.2
django-crispy-forms==2.1
django-environ==0.11.2
django-htmx==1.17.2
django-redis==5.4.0
gunicorn==21.2.0
humanfriendly==10.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
msgpack==1.1.0
packaging==23.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==24.0.0
python-http-client==3.3.7
redis==5.0.1
requests==2.31.0
semver==3.0.2
sendgrid==6.11.0
service-identity==24.1.0
setuptools==69.0.3
six==1.16.0
sqlparse==0.4.4
starkbank-ecdsa==2.2.0
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.12.0
urllib3==2.2.0
whitenoise==6.6.0
zope.interface==7.1.1
djangorestframework==3.15.2
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-api-python-client==2.108.0
pdfminer.six==20251107
pdfplumber>=0.10.3
pandas>=2.2.0
celery==5.3.4
celery-progress==0.5
pytesseract
google-generativeai==0.3.2
openai>=1.0.0
anthropic>=0.18.0
drf-spectacular==0.27.1
mcp
django-vite
django-htmx==1.17.2
django-filter
django-cors-headers
channels
channels-redis
daphne
